refactor: export public utils
