+++ .changeset/fuzzy-needles-bow.md
@@ -1,0 +1,10 @@
+---
+'@floating-ui/utils': minor
+'@floating-ui/core': patch
+'@floating-ui/dom': patch
+'@floating-ui/react': patch
+'@floating-ui/react-dom': patch
+'@floating-ui/vue': patch
+---
+
+refactor: promote stable utils
