+++ packages/core/src/middleware/arrow.ts
@@ -1,13 +1,11 @@
 import {
   clamp,
   evaluate,
-  getAlignment,
-  getAlignmentAxis,
-  getAxisLength,
   getPaddingObject,
   min as mathMin,
 } from '@floating-ui/utils';
 
+import {getAlignment, getAlignmentAxis, getAxisLength} from '../';
 import type {Derivable, Middleware, Padding} from '../types';
 
 export interface ArrowOptions {
