+++ packages/dom/src/platform/getElementRects.ts
@@ -1,4 +1,4 @@
-import {Platform} from '../types';
+import type {Platform} from '../types';
 import {getRectRelativeToOffsetParent} from '../utils/getRectRelativeToOffsetParent';
 import {getOffsetParent} from './getOffsetParent';
