+++ packages/dom/src/types.ts
@@ -286,5 +286,19 @@ export type {
   SideObject,
   Strategy,
 } from '@floating-ui/core';
-export {offset} from '@floating-ui/core';
+export {
+  alignments,
+  getAlignment,
+  getAlignmentAxis,
+  getAxisLength,
+  getOppositeAlignmentPlacement,
+  getOppositeAxis,
+  getOppositeSidePlacement,
+  getSide,
+  getSideAxis,
+  offset,
+  placements,
+  rectToClientRect,
+  sides,
+} from '@floating-ui/core';
 export {getOverflowAncestors} from '@floating-ui/utils/dom';
