+++ packages/react-dom/src/index.ts
@@ -15,3 +15,17 @@ export {
   shift,
   size,
 } from '@floating-ui/dom';
+export {
+  alignments,
+  getAlignment,
+  getAlignmentAxis,
+  getAxisLength,
+  getOppositeAlignmentPlacement,
+  getOppositeAxis,
+  getOppositeSidePlacement,
+  getSide,
+  getSideAxis,
+  placements,
+  rectToClientRect,
+  sides,
+} from '@floating-ui/dom';
