+++ packages/react-dom/src/types.ts
@@ -63,6 +63,20 @@ export {
   shift,
   size,
 } from '@floating-ui/dom';
+export {
+  alignments,
+  getAlignment,
+  getAlignmentAxis,
+  getAxisLength,
+  getOppositeAlignmentPlacement,
+  getOppositeAxis,
+  getOppositeSidePlacement,
+  getSide,
+  getSideAxis,
+  placements,
+  rectToClientRect,
+  sides,
+} from '@floating-ui/dom';
 
 type Prettify<T> = {
   [K in keyof T]: T[K];
