+++ packages/react/src/index.ts
@@ -50,3 +50,17 @@ export {
   shift,
   size,
 } from '@floating-ui/react-dom';
+export {
+  alignments,
+  getAlignment,
+  getAlignmentAxis,
+  getAxisLength,
+  getOppositeAlignmentPlacement,
+  getOppositeAxis,
+  getOppositeSidePlacement,
+  getSide,
+  getSideAxis,
+  placements,
+  rectToClientRect,
+  sides,
+} from '@floating-ui/react-dom';
