+++ packages/react/src/types.ts
@@ -81,6 +81,20 @@ export {
   shift,
   size,
 } from '@floating-ui/react-dom';
+export {
+  alignments,
+  getAlignment,
+  getAlignmentAxis,
+  getAxisLength,
+  getOppositeAlignmentPlacement,
+  getOppositeAxis,
+  getOppositeSidePlacement,
+  getSide,
+  getSideAxis,
+  placements,
+  rectToClientRect,
+  sides,
+} from '@floating-ui/react-dom';
 
 type Prettify<T> = {
   [K in keyof T]: T[K];
