--- packages/utils/react/src/types.ts
@@ -1,1 +1,0 @@
-export * from './';
