+++ packages/vue/src/types.ts
@@ -67,6 +67,20 @@ export {
   shift,
   size,
 } from '@floating-ui/dom';
+export {
+  alignments,
+  getAlignment,
+  getAlignmentAxis,
+  getAxisLength,
+  getOppositeAlignmentPlacement,
+  getOppositeAxis,
+  getOppositeSidePlacement,
+  getSide,
+  getSideAxis,
+  placements,
+  rectToClientRect,
+  sides,
+} from '@floating-ui/dom';
 
 export type MaybeReadonlyRef<T> = T | Readonly<Ref<T>>;
